+++ app/Resources/views/base.html.twig
@@ -78,19 +78,7 @@
                             {% endif %}
                         </ul>
                     </li>
-                    <li>
-                        {{ form_start(form, {'attr': {'class': 'navbar-form navbar-right'}}) }}
-                        {{ form_errors(form) }}
-
-                        <div class="form-group">
-                            {{ form_widget(form.recherche, {'attr': {'class': 'form-control', 'placeholder': 'Recherche'}})}}
-                            {{ form_errors(form.recherche) }}
-                        </div>
-                        <button type="submit" id="btn_search" class="btn"><span class="glyphicon glyphicon-search"
-                                                                                aria-hidden="true"></span></button>
-                        {{ form_rest(form) }}
-                        {{ form_end(form) }}
-                    </li>
+                    {{  render(controller('AppBundle:Recherche:formRecherche')) }}
                 </ul>
             </div>
         </div>
@@ -118,6 +106,43 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
     <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
     <script src="{{ asset('js/recherche.js')}}"></script>
+    <script>
+        $(function(){
+            $('#recherche_recherche').autocomplete({
+                source : function (requete, reponse) { // la fonction anonyme permet de maintenir une requête AJAX directement dans le plugin
+                    //un qui représentera la requête tapée dans le champ de texte, et un autre qui représentera la réponse renvoyée par AJAX, et donc ma liste de suggestions.
+                    var objData = {};
+                    objData = { nomVern: requete.term };
+                    console.log(objData);
+                    $.ajax({
+                        url : "{{ path('ajax') }}", //On appelle le script JSON
+                        dataType : 'json', //On spécifie bien que le type de donnée est en JSON
+                        data : objData,
+                        type: "POST",
+                        success : function(data){  // on se prépare à renvoyer les données réceptionnées grâce à l'évènement de succès
+                            //console.log(data);
+                            reponse($.map(data, function(objet){
+                                //var objet[] = objet.title + objet.nomVern;
+                                console.log(objet);
+                                if(objet.title == null)
+                                {
+                                    return objet.nomVern;
+                                }
+                                if(objet.nomVern == null)
+                                {
+                                    return objet.title;
+                                }
+                                //return objet.nomVern;// on retourne cette forme de suggestion
+                            }))
+                        },
+                        error: function() { // if error
+                            alert('problème');
+                        }
+                    });
+                }
+            });
+        });
+    </script>
     {% block JS %} {% endblock %}
 {% endblock %}
 </body>
